Separate client surfaces out into game and puzzle
