+++ jigsaw/client/__init__.py
@@ -1,2 +1,2 @@
 #!/usr/bin/venv python3
-from .gui import GameClient
+from .game import GameClient
